<Feat> : File IO Optimazation works right.
Converting text to binary works right.
Script size shrinked almost 1/4 ~ 1/5.

+++ Project/KGCA41/DirectXBasic/Mesh.cpp
@@ -711,10 +711,10 @@ namespace SSB
 					delete tmpBuffer;
 				}
 				{
-					int tmpSize = vertex.Weight[i];
-					char* tmpBuffer = new char[sizeof(int)];
-				memcpy(tmpBuffer, &tmpSize, sizeof(int));
-					std::string tmpStr(tmpBuffer, sizeof(int));
+					float tmpSize = vertex.Weight[i];
+					char* tmpBuffer = new char[sizeof(float)];
+				memcpy(tmpBuffer, &tmpSize, sizeof(float));
+					std::string tmpStr(tmpBuffer, sizeof(float));
 					ret += tmpStr;
 					delete tmpBuffer;
 				}
